+++ src/commands/configCommands.ts
@@ -1,0 +1,9 @@
+import * as vscode from 'vscode';
+
+export async function openConfig() {
+    
+}
+
+export function initConfigCmd(context: vscode.ExtensionContext) {
+    context.subscriptions.push(vscode.commands.registerCommand('slurm--.openConfig', openConfig));
+}