+++ src/commands/resourceCommand.ts
@@ -33,7 +33,7 @@ function extractNodes(nodeString: string, short_length: number, long_length: num
     return nodeList;
 }
 
-export async function refreshResources() {
+async function refreshResources() {
     const short = 15;
     const long = 50;
     const [out, err] = await executeCmd(`sinfo --noheader --Node -O NODELIST:${short},Available:${short},Memory:${short},AllocMem:${short},Gres:${long},GresUsed:${long},Partition:${short},StateLong:${short}`);
@@ -41,7 +41,7 @@ export async function refreshResources() {
     resourceViewDataProvider.refresh();
 }
 
-export async function autoRefreshRes() {
+async function autoRefreshRes() {
     clearInterval(autoRefreshTimer);
     autoRefreshTimer = setInterval(() => {
         vscode.commands.executeCommand('slurm--.refreshResources');
@@ -49,12 +49,12 @@ export async function autoRefreshRes() {
     vscode.commands.executeCommand('setContext', 'autoRefreshingRes', true);
 }
 
-export async function unautoRefreshRes() {
+async function unautoRefreshRes() {
     clearInterval(autoRefreshTimer);
     vscode.commands.executeCommand('setContext', 'autoRefreshingRes', false);
 }
 
-export async function copyGres() {
+async function copyGres() {
     const selected = resourceTreeView.selection[0];
     if (selected instanceof ListItem){
         const gresString = (selected.title as string).replace(/\s*\(.*\)$/, '');