+++ src/models/nodeModel.ts
@@ -12,8 +12,9 @@ export class Node {
     readonly memory: number;
     allocMemory: number;
     readonly gres: ResourceGres | null;
-    state: NodeState;
-    //NODELIST:15,Available:15,Memory:15,AllocMem:15,Gres:50,GresUsed:50,Partition:15
+    readonly partition: string;
+    readonly state: NodeState;
+    //NODELIST:15,Available:15,Memory:15,AllocMem:15,Gres:50,GresUsed:50,Partition:15,StateLong:15
     constructor(nodeid: string, available: string, memory: string, allocMemory: string, gres: string, usedGres: string, partition: string, state: string)
 
     constructor(
@@ -23,12 +24,13 @@ export class Node {
         allocMemory: string,
         gres: string,
         usedGres: string,
-        readonly partition: string,
+        partition: string,
         state: string,
     ) {
         this.memory = (typeof memory === "string" ? parseInt(memory) : memory) / 1000;
         this.allocMemory = (typeof allocMemory === "string" ? parseInt(allocMemory) : allocMemory) / 1000;
         this.gres = gres === '(null)' ? null : new ResourceGres(usedGres, gres);
+        this.partition = partition.endsWith('*') ? partition.slice(0, partition.length - 1) : partition;
         this.state = (state.endsWith('*') ? state.slice(0, state.length - 1) : state) as NodeState;
     }